+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 
 # Celestra version history
 
+## Celestra v7.1.0 Scotty
+
+1. Documentation, pdf and code fixes.
+2. Remove _(stability 0)_ these functions:
+
+- `isInt8(value);`
+- `isInt16(value);`
+- `isInt32(value);`
+- `isUInt32(value);`
+- `isUInt8(value);`
+- `isUInt16(value);`
+- `isBigInt64(value);`
+- `isBigUInt64(value);`
+- `isFloat16(value);`
+- `nanoid();`
+- `toInt8(value);`
+- `toInt16(value);`
+- `toInt32(value);`
+- `toUInt8(value);`
+- `toUInt16(value);`
+- `toUInt32(value;`
+- `toBigInt64(value);`
+- `toBigUInt64(value);`
+- `toFloat16(value);`
+- `toFloat32(value);`
+
 ## Celestra v7.0.1
 
 1. Documentation, pdf and code fixes.